update_7

+++ app/src/main/java/com/sbdevs/bookonline/adapters/ProductReviewAdapter.kt
@@ -43,7 +43,7 @@ class ProductReviewAdapter(var list: MutableList<ProductReviewModel>) :
 
             val buyerId: String = reviewModel.buyer_ID
             val buyerName: String = reviewModel.buyer_name
-            val rating: Int = reviewModel.rating
+            val rating: Long = reviewModel.rating
             val review: String = reviewModel.review
             val reviewDate = reviewModel.review_Date!!
             val daysAgo = TimeDateAgo().msToTimeAgo(itemView.context,reviewDate)
